spent 3hrs reseaching further into custom editors and property drawer.
tried making custom draw property for AttackData class i cant seem to get it to work or make it look any different in the inspector.

tried finding a way that allows me to view AttackData scriptable objects elements in moveSetData list from the EntityData scriptable object and edit them, but couldnt find a way to get the elements to show.