spent 2hrs 30mins working out how to build maps and place map nodes at runtime, got a simple example up and working. and fixed some ui issues in scene.
moves map node related things out of SpawnManager for now and into InteractiveMapHandler.

updated InteractiveMapHandler script:
added summery of a simple map layout i want to make.
added func to set up and generate a map, a func to generate each column and add nodes in the same column to a dictionary + initilize the map node. added logic to handle weighted map node spawn chance and return it.
functions to set map nodes postions on the map based on number of columns + number of nodes in said column.

TODO:
refactor code to be less of a mess when generation map nodes and setting there positions + add links between them to unlock the next map node once encounter is won.