spent 3hrs updating code to use package AbilitySystem versions of scripts. + creating nes ComplexStats package to try and fix issues. currently looking into ways to still fix them. made little progress to getting it working without just having an enum i endlessly add to.
TODO: figure out a way (if there is any) to be able to create a list of stat types per game and have the ComplexStats package be able to use and accepth them through generic types.
atm considering just having an enum of stat types be defined outside of the package. (paartially have the code set up if it does work)