spent 45mins experimenting with what stat types and status effects i could add. ran into issues with complex stat package not really being able to support what i want.
TODO:
update complex stat package to support a stat being able to handle float and int modifiers. eg: attack dmg = 5 + 2 dmg * -0.25 dmg for a final value.
pprobably inclued a StatValue class that is just a simple bool + float value. to make it easer figuing out if a value is supposed to be treated as a percentage or not.